pyproject.toml
[tool.poetry]
name = "maya"
version = "0.1.0"
description = ""
authors = ["cacko <alex@cacko.net>"]

[[tool.poetry.source]]
name = "cacko"
secondary = true
url = "https://pypi.cacko.net/simple/"


[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.2"
humanfriendly = "^10.0"
boto3 = "^1.21.32"
Pillow = "^9.1.0"
dataclasses-json = "^0.5.7"
exif = "^1.3.4"
tqdm = "^4.63.1"
google-cloud-firestore = "^2.4.0"
filetype = "^1.0.10"
peewee = "^3.14.10"
psycopg2 = "^2.9.3"
Flask = "^2.1.1"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
